Add readOnly prop (#1316)
* Add readOnly prop

* Update e2e/integration/DatePicker.spec.ts
